requirements.txt
fastapi
uvicorn
sqlalchemy
psycopg2-binary
python-dotenv==1.0.0